// src/app/database/entities/financiero/cuentas-por-pagar-enums.ts
export enum CuotaEstado {
  PENDIENTE = 'PENDIENTE',
  PARCIAL = 'PARCIAL',
  PAGADA = 'PAGADA',
  VENCIDA = 'VENCIDA'
}

export enum CuentaPorPagarEstado {
  ACTIVO = 'ACTIVO',
  PAGADO = 'PAGADO',
  CANCELADO = 'CANCELADO'
}

export enum CuentaPorPagarTipo {
  COMPRA = 'COMPRA',
  PRESTAMO = 'PRESTAMO',
  OTRO = 'OTRO'
}