// src/app/database/entities/financiero/caja-mayor-enums.ts
export enum CajaMayorEstado {
  ABIERTA = 'ABIERTA',
  CERRADA = 'CERRADA'
}

export enum TipoMovimiento {
  INGRESO_RETIRO_CAJA = 'INGRESO_RETIRO_CAJA',
  INGRESO_CIERRE_CAJA = 'INGRESO_CIERRE_CAJA',
  INGRESO_ENTRADA_VARIA = 'INGRESO_ENTRADA_VARIA',
  INGRESO_OPERACION_FINANCIERA = 'INGRESO_OPERACION_FINANCIERA',
  INGRESO_RETIRO_BANCO = 'INGRESO_RETIRO_BANCO',
  TRANSFERENCIA_ENTRADA = 'TRANSFERENCIA_ENTRADA',
  AJUSTE_POSITIVO = 'AJUSTE_POSITIVO',
  EGRESO_GASTO = 'EGRESO_GASTO',
  EGRESO_COMPRA = 'EGRESO_COMPRA',
  EGRESO_CUOTA_COMPRA = 'EGRESO_CUOTA_COMPRA',
  EGRESO_CUOTA_PRESTAMO = 'EGRESO_CUOTA_PRESTAMO',
  EGRESO_VALE = 'EGRESO_VALE',
  EGRESO_SALARIO = 'EGRESO_SALARIO',
  EGRESO_CHEQUE = 'EGRESO_CHEQUE',
  EGRESO_OPERACION_FINANCIERA = 'EGRESO_OPERACION_FINANCIERA',
  EGRESO_DEPOSITO_BANCO = 'EGRESO_DEPOSITO_BANCO',
  EGRESO_CAJA_INICIAL = 'EGRESO_CAJA_INICIAL',
  TRANSFERENCIA_SALIDA = 'TRANSFERENCIA_SALIDA',
  AJUSTE_NEGATIVO = 'AJUSTE_NEGATIVO',
  ANULACION = 'ANULACION'
}

export enum GastoEstado {
  PENDIENTE = 'PENDIENTE',
  PAGADO = 'PAGADO',
  PROGRAMADO = 'PROGRAMADO',
  CANCELADO = 'CANCELADO'
}

export enum GastoFrecuencia {
  DIARIO = 'DIARIO',
  SEMANAL = 'SEMANAL',
  QUINCENAL = 'QUINCENAL',
  MENSUAL = 'MENSUAL',
  BIMESTRAL = 'BIMESTRAL',
  TRIMESTRAL = 'TRIMESTRAL',
  SEMESTRAL = 'SEMESTRAL',
  ANUAL = 'ANUAL'
}

export enum RetiroCajaEstado {
  FLOTANTE = 'FLOTANTE',
  VINCULADO_PENDIENTE = 'VINCULADO_PENDIENTE',
  INGRESADO = 'INGRESADO'
}